requirements.txt
Django==4.2
djangorestframework==3.16.1
requests==2.31.0
python-dotenv==1.0.1
python-decouple==3.8
bcrypt==3.2.2
cryptography==41.0.7
PyYAML==6.0.1
gunicorn==21.2.0
django-cors-headers==4.3.0
dj-rest-auth==7.0.1
django-allauth==0.57.0
python-slugify==8.0.1
djangorestframework-simplejwt==5.5.1
django-filter==23.2
psycopg2-binary==2.9.9
pytz>=2025.2
Pillow>=10.0.0  # Para procesamiento de imágenes (firmas)
PyPDF2>=3.0.0  # Para trabajar con PDFs (futuro)
reportlab>=4.0.0  # Para generar PDFs con firmas (futuro)